refactor: Move GuDeveloperPolicy out of experiment
