+++ src/constructs/iam/policies/developer-policy.test.ts
@@ -1,12 +1,12 @@
 import { Annotations, Template } from "aws-cdk-lib/assertions";
 import { Effect, PolicyStatement } from "aws-cdk-lib/aws-iam";
-import { simpleGuStackForTesting } from "../../../../utils/test";
-import { GuDeveloperPolicyExperimental } from "./developer-policy";
+import { simpleGuStackForTesting } from "../../../utils/test";
+import { GuDeveloperPolicy } from "./developer-policy";
 
-describe("GuDeveloperPolicyExperimental", () => {
-  test("if a single policy statement is provided, the resulting Workload Policy resource's statement will have a single item", () => {
+describe("GuDeveloperPolicy", () => {
+  test("if a single policy statement is provided, the resulting Developer Policy resource's statement will have a single item", () => {
     const stack = simpleGuStackForTesting();
-    new GuDeveloperPolicyExperimental(stack, "AllowS3GetObject", {
+    new GuDeveloperPolicy(stack, "AllowS3GetObject", {
       statements: [
         PolicyStatement.fromJson({
           Action: ["s3:GetObject"],
@@ -34,7 +34,7 @@ describe("GuDeveloperPolicyExperimental", () => {
 
   test("adds an error if a wide-open resource/allow policy statement is present", () => {
     const stack = simpleGuStackForTesting();
-    new GuDeveloperPolicyExperimental(stack, "AllowS3GetObject", {
+    new GuDeveloperPolicy(stack, "AllowS3GetObject", {
       statements: [
         PolicyStatement.fromJson({
           Effect: "Allow",
@@ -49,7 +49,7 @@ describe("GuDeveloperPolicyExperimental", () => {
 
   test("adds an error if a wide-open resource/allow policy statement is present", () => {
     const stack = simpleGuStackForTesting();
-    new GuDeveloperPolicyExperimental(stack, "AllowS3GetObject", {
+    new GuDeveloperPolicy(stack, "AllowS3GetObject", {
       statements: [
         PolicyStatement.fromJson({
           Effect: "Allow",
@@ -64,7 +64,7 @@ describe("GuDeveloperPolicyExperimental", () => {
 
   test("adds an error if a wide-open policy statement with no effect is present", () => {
     const stack = simpleGuStackForTesting();
-    new GuDeveloperPolicyExperimental(stack, "AllowS3GetObject", {
+    new GuDeveloperPolicy(stack, "AllowS3GetObject", {
       statements: [
         PolicyStatement.fromJson({
           Action: ["*"],
@@ -78,7 +78,7 @@ describe("GuDeveloperPolicyExperimental", () => {
 
   test("does not add an error if a wide-open policy statement with a Deny effect is present", () => {
     const stack = simpleGuStackForTesting();
-    new GuDeveloperPolicyExperimental(stack, "AllowS3GetObject", {
+    new GuDeveloperPolicy(stack, "AllowS3GetObject", {
       statements: [
         PolicyStatement.fromJson({
           Effect: "Deny",
@@ -93,7 +93,7 @@ describe("GuDeveloperPolicyExperimental", () => {
 
   test("adds an error if a wide-open action/allow policy statement is present", () => {
     const stack = simpleGuStackForTesting();
-    new GuDeveloperPolicyExperimental(stack, "AllowS3GetObject", {
+    new GuDeveloperPolicy(stack, "AllowS3GetObject", {
       statements: [
         PolicyStatement.fromJson({
           Effect: "Allow",
@@ -108,7 +108,7 @@ describe("GuDeveloperPolicyExperimental", () => {
 
   test("adds an error if a wide-open statement resource is requested", () => {
     const stack = simpleGuStackForTesting();
-    new GuDeveloperPolicyExperimental(stack, "AllowS3GetObject", {
+    new GuDeveloperPolicy(stack, "AllowS3GetObject", {
       statements: [
         new PolicyStatement({
           effect: Effect.ALLOW,
@@ -123,7 +123,7 @@ describe("GuDeveloperPolicyExperimental", () => {
 
   test("adds an error if a wide-open statement action is requested", () => {
     const stack = simpleGuStackForTesting();
-    new GuDeveloperPolicyExperimental(stack, "AllowS3GetObject", {
+    new GuDeveloperPolicy(stack, "AllowS3GetObject", {
       statements: [
         new PolicyStatement({
           effect: Effect.ALLOW,
@@ -138,7 +138,7 @@ describe("GuDeveloperPolicyExperimental", () => {
 
   test("adds an error if both a wide-open statement action and a resource is requested", () => {
     const stack = simpleGuStackForTesting();
-    new GuDeveloperPolicyExperimental(stack, "AllowS3GetObject", {
+    new GuDeveloperPolicy(stack, "AllowS3GetObject", {
       statements: [
         new PolicyStatement({
           effect: Effect.ALLOW,
