--- src/experimental/constructs/iam/policies/index.ts
@@ -1,1 +1,0 @@
-export * from "./developer-policy";
